+++ libft/ft_memchr.c
@@ -6,14 +6,16 @@
 /*   By: mcolomba <mcolomba@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/23 14:19:17 by mcolomba          #+#    #+#             */
-/*   Updated: 2022/10/24 12:50:28 by mcolomba         ###   ########.fr       */
+/*   Updated: 2022/11/03 11:20:13 by mcolomba         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memchr(const void *s, int c, unsigned int n)
+#include "libft.h"
+
+void	*ft_memchr(const void *s, int c, size_t n)
 {
-	const char		*pos;
-	unsigned int	i;
+	const char	*pos;
+	size_t		i;
 
 	pos = (const char *) s;
 	i = 0;
