+++ libft/ft_memset.c
@@ -6,11 +6,13 @@
 /*   By: mcolomba <mcolomba@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/21 18:39:32 by marcos            #+#    #+#             */
-/*   Updated: 2022/09/22 13:22:34 by mcolomba         ###   ########.fr       */
+/*   Updated: 2022/11/03 11:15:59 by mcolomba         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *s, int c, unsigned int n)
+#include "libft.h"
+
+void	*ft_memset(void *s, int c, size_t n)
 {
 	unsigned char	*aux;
